fix: widen multiplication before assigning to size_t (CodeQL alerts #21/#22)
d3d9_texture.h's subresource-count calc and hud_renderer.cpp's font
texture-data-size calc both multiplied two 32-bit values before letting
the result implicitly widen to size_t -- a real, if low-practical-risk,
truncate-then-widen overflow pattern. Cast each operand to size_t before
multiplying instead.

+++ dxvk/src/d3d9/d3d9_texture.h
@@ -32,7 +32,7 @@ namespace dxvk {
       const uint32_t arraySlices = m_texture.Desc()->ArraySize;
       const uint32_t mipLevels   = m_texture.Desc()->MipLevels;
 
-      m_subresources.resize(arraySlices * mipLevels);
+      m_subresources.resize(size_t(arraySlices) * size_t(mipLevels));
 
       for (uint32_t i = 0; i < arraySlices; i++) {
         for (uint32_t j = 0; j < mipLevels; j++) {
